% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghfd_lob_fcts.R
\name{add.order}
\alias{add.order}
\title{Adds an order to the LOB}
\usage{
add.order(my.lob, order.in, silent = TRUE)
}
\arguments{
\item{my.lob}{A LOB (order book)}

\item{order.in}{An order from the data}

\item{silent}{Should the function print progress ? (TRUE (default) or FALSE)}
}
\value{
An LOB with the new order
}
\description{
Adds an order to the LOB
}
\examples{
# no example (internal)
}
