% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghfd_utils.R
\name{get.info.opt}
\alias{get.info.opt}
\title{Extracts information from options using ticker symbol
#'}
\usage{
get.info.opt(ticker.in, session.date)
}
\arguments{
\item{ticker.in}{The ticker of the options (e.f. GGBRB15)}

\item{session.date}{The date of the trading session  (e.g. as.Date('2010-01-01'))}
}
\value{
A list with the following items: \describe{
\item{type.option}{The type of options (CALL or PUT)}
\item{strike.price}{The strike price of option}
\item{maturity.date}{The maturity date of option (see <http://www.bmfbovespa.com.br/pt_br/produtos/listados-a-vista-e-derivativos/renda-variavel/opcoes-sobre-acoes.htm>)} }
}
\description{
Extracts information from options using ticker symbol
#'
}
\examples{

info <- get.info.opt('GGBRA18', as.Date('2010-01-01'))
}
