% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghfd_utils.R
\name{GetNWeekDayOfMonth}
\alias{GetNWeekDayOfMonth}
\title{Find the N weekday of a month/year}
\usage{
GetNWeekDayOfMonth(my.month, my.year, my.weekday = "Monday", my.pos = 3)
}
\arguments{
\item{my.month}{The month, as numeric (e.g. 02)}

\item{my.year}{The year, as numeric (e.g. 2010)}

\item{my.weekday}{The weekday, not abbreviated (e.g. 'Monday')}

\item{my.pos}{The position of weekday within the month, as numeric (e.g. third monday of month: my.pos = 3)}
}
\value{
Returns a Date object
}
\description{
Find the N weekday of a month/year
}
\examples{

date.out <-  GetNWeekDayOfMonth(my.month = 02, my.year = 2010, my.weekday = 'Monday', my.pos = 3)
}
