% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghfd_read_file.R
\name{ghfd_read_file.orders}
\alias{ghfd_read_file.orders}
\title{Reads zip file downloaded from Bovespa ftp (orders) - INTERNAL USE}
\usage{
ghfd_read_file.orders(out.file, my.assets = NULL, type.matching,
  first.time = "10:00:00", last.time = "17:00:00", type.output = "agg",
  agg.diff = "15 min")
}
\arguments{
\item{out.file}{Name of zip file}

\item{my.assets}{The tickers (symbols) of the derised assets to import data (e.g. c('PETR4', 'VALE5')). The function allow for partial patching (e.g. 'PETR' for all assets related to Petrobras). Default is set to NULL (download all available tickers)}

\item{type.matching}{Type of matching for asset names in data ('exact' or 'partial')}

\item{first.time}{The first intraday period to import the data. All trades/orders before this time of day are ignored. As character, e.g. '10:00:00'.}

\item{last.time}{The last intraday period to import the data. All trades/orders after this time of day are ignored. As character, e.g. '18:00:00'.}

\item{type.output}{Defines the type of output of the data. The choice 'agg' outputs aggregated data for time intervals defined in agg.diff.
The choice 'raw' outputs the raw, tick by tick/order by order, data from the zip files.}

\item{agg.diff}{The time interval used in the aggregation of data. Only used for type.output='agg'. It should contain a integer followed by a time unit ('sec' or 'secs', 'min' or 'mins', 'hour' or 'hours', 'day' or 'days').
Example: agg.diff = '15 mins', agg.diff = '1 hour'.}
}
\value{
A dataframe with trade data (aggregated or raw)
}
\description{
Reads zip file downloaded from Bovespa ftp (orders) - INTERNAL USE
}
\examples{

# no example
}
