% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghfd_lob_fcts.R
\name{organize.lob}
\alias{organize.lob}
\title{Organizes LOB (internal function)}
\usage{
organize.lob(my.lob, silent = TRUE)
}
\arguments{
\item{my.lob}{A LOB (order book)}

\item{silent}{Should the function print progress ? (TRUE (default) or FALSE)}
}
\value{
An organized LOB
}
\description{
This internal recursive function organizes the lob by making sure that all prices and time are ordered.
Every time that prices in the bid and ask matches, it will create a trade and modify the lob accordingly.
}
\examples{

# no examples (internal)
}
