% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghfd_lob_fcts.R
\name{print.lob}
\alias{print.lob}
\title{Prints the LOB}
\usage{
\method{print}{lob}(my.lob, max.level = 3)
}
\arguments{
\item{my.lob}{A LOB (order book)}

\item{max.level}{Max level of lob to print}
}
\value{
nothing
}
\description{
Prints the LOB
}
\examples{
# no example (internal)
}
