% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghfd_lob_fcts.R
\name{process.lob.from.df}
\alias{process.lob.from.df}
\title{Process LOB from asset dataframe}
\usage{
process.lob.from.df(asset.df, silent = T)
}
\arguments{
\item{asset.df}{A dataframe with orders for a single asset}

\item{silent}{Should the function print progress ? (TRUE (default) or FALSE)}
}
\value{
The lob for the single asset
}
\description{
Process LOB from asset dataframe
}
\examples{
# no example (internal)
}
